Move TranslationModal to FieldGroupSettings panel
